Refactor internal package structure
