+++ src/neptune_scale/api/__init__.py
@@ -1,0 +1,1 @@
+# This package contains user-facing classes and functions.
