+++ src/neptune_scale/net/__init__.py
@@ -1,0 +1,1 @@
+# This package contains everything related to communication with the Neptune API.
