+++ src/neptune_scale/net/serialization.py
