+++ src/neptune_scale/sync/errors_tracking.py
@@ -11,11 +11,7 @@ from typing import (
     Type,
 )
 
-from neptune_scale.core.components.abstract import Resource
-from neptune_scale.core.components.daemon import Daemon
-from neptune_scale.core.logger import get_logger
-from neptune_scale.core.process_killer import kill_me
-from neptune_scale.exceptions import (
+from neptune_scale.api.exceptions import (
     NeptuneAsyncLagThresholdExceeded,
     NeptuneConnectionLostError,
     NeptuneOperationsQueueMaxSizeExceeded,
@@ -23,7 +19,11 @@ from neptune_scale.exceptions import (
     NeptuneScaleWarning,
     NeptuneUnexpectedError,
 )
-from neptune_scale.parameters import ERRORS_MONITOR_THREAD_SLEEP_TIME
+from neptune_scale.sync.parameters import ERRORS_MONITOR_THREAD_SLEEP_TIME
+from neptune_scale.util import get_logger
+from neptune_scale.util.abstract import Resource
+from neptune_scale.util.daemon import Daemon
+from neptune_scale.util.process_killer import kill_me
 
 logger = get_logger()
