+++ src/neptune_scale/sync/metadata_splitter.py
@@ -25,7 +25,7 @@ from neptune_api.proto.neptune_pb.ingest.v1.common_pb2 import (
 )
 from neptune_api.proto.neptune_pb.ingest.v1.pub.ingest_pb2 import RunOperation
 
-from neptune_scale.core.serialization import (
+from neptune_scale.net.serialization import (
     datetime_to_proto,
     make_step,
     make_value,
