+++ src/neptune_scale/util/abstract.py
