+++ src/neptune_scale/util/shared_var.py
