+++ tests/unit/test_operations_queue.py
@@ -7,7 +7,7 @@ from neptune_api.proto.neptune_pb.ingest.v1.common_pb2 import (
 )
 from neptune_api.proto.neptune_pb.ingest.v1.pub.ingest_pb2 import RunOperation
 
-from neptune_scale.core.components.operations_queue import OperationsQueue
+from neptune_scale.sync.operations_queue import OperationsQueue
 
 
 def test__enqueue():
